+++ src/lib/options-repository.ts
@@ -1,4 +1,5 @@
-import { Repository, RepositoryProps } from './repository'
+import { RepositoryProps } from '../types'
+import { Repository } from './repository'
 
 export class OptionsRepository extends Repository {
   constructor(props: RepositoryProps) {
