+++ src/lib/repository.ts
@@ -1,9 +1,6 @@
-import { BrowserStorage, Browser, StorageArea } from './browser-storage'
+import { RepositoryProps } from '../types'
+import { BrowserStorage } from './browser-storage'
 
-export type RepositoryProps = {
-  browser: Browser
-  storageArea: StorageArea
-}
 export class Repository {
   storage: BrowserStorage<string>
   key: string
